Add files via upload
The model is for both public and government purposes.
Our model is like a smart friend for road safety and environment. It watches out for blind spots on the road to prevent accidents – those tricky areas where drivers can't see. It also checks the air quality, letting us know if the air is clean or if there's pollution. When it comes to the weather, it predicts what it's going to be like so you can plan your day. And for areas where they might need better water routes (like canals or rivers), it predicts if there is a possibility for waterway development. The model also prompts the user(government admin only) when the road needs to be repaired or developed. It's like having a helpful buddy for road safety and the environment!